+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ https://gitlab.com/gf-bl/oereb_client/milestones/16
 
 - Adjust legend titles according to PDF specification
 - Add responsible offices from geometries
+- Specify OEREB logo via configuration
 
 1.2.3
 *****